+++ eng/scripts/Get-PkgVersion.ps1
@@ -7,26 +7,36 @@ param (
     [string] $PackageName
 )
 
-. ${PSScriptRoot}/SdkVersion-Common.ps1
+$repoRoot = Resolve-Path "$PSScriptRoot/../..";
+. (Join-Path ${repoRoot} eng common scripts logging.ps1)
+. (Join-Path ${repoRoot} eng scripts SdkVersion-Common.ps1)
 
 $versionFileLocation = Get-VersionHppLocaiton `
     -ServiceDirectory $ServiceDirectory `
     -PackageName $PackageName
 
 if (!$versionFileLocation) {
-    $fallback = Get-Content $RepoRoot/sdk/$ServiceDirectory/$PackageName/version.txt
+    $fallbackpath = Join-Path $RepoRoot sdk $ServiceDirectory $PackageName version.txt
+    if (!(Test-Path $fallbackpath))
+    {
+        LogWarning "Failed to retrieve package version. No version file found."
+        return $null
+    }
+
+    $fallback = Get-Content $fallbackpath
     if ($fallback) {
         return $fallback
     } else {
-        Write-Error "Cannot locate package version"
-        exit 1
+        LogWarning "Cannot locate package version"
+        return $null
     }
 }
 
 $versionFileContents = Get-Content $versionFileLocation -Raw
 
 if (!($versionFileContents -match $VersionRegex)) {
-    Write-Error "does not match version information schema"
+    LogWarning "does not match version information schema"
+    return $null
 }
 
 $VersionString = if ($Matches.prerelease) {