+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client.cpp
@@ -23,28 +23,20 @@ KeyClient::KeyClient(
 {
   auto apiVersion = options.GetVersionString();
 
-  // Base Pipeline
-  std::vector<std::unique_ptr<HttpPolicy>> policies;
-  policies.emplace_back(
-      std::make_unique<TelemetryPolicy>("KeyVault", apiVersion, options.TelemetryPolicyOptions));
-  policies.emplace_back(std::make_unique<RequestIdPolicy>());
-  policies.emplace_back(std::make_unique<RetryPolicy>(options.RetryOptions));
-
+  std::vector<std::unique_ptr<HttpPolicy>> perRetrypolicies;
   {
     Azure::Core::Http::TokenRequestOptions const tokenOptions
         = {{"https://vault.azure.net/.default"}};
 
-    policies.emplace_back(
+    perRetrypolicies.emplace_back(
         std::make_unique<BearerTokenAuthenticationPolicy>(credential, tokenOptions));
   }
 
-  policies.emplace_back(std::make_unique<LoggingPolicy>());
-  policies.emplace_back(
-      std::make_unique<Azure::Core::Http::TransportPolicy>(options.TransportPolicyOptions));
-  Azure::Core::Http::Url url(vaultUrl);
-
   m_pipeline = std::make_shared<Azure::Security::KeyVault::Common::Internal::KeyVaultPipeline>(
-      url, apiVersion, std::move(policies));
+      Azure::Core::Http::Url(vaultUrl),
+      apiVersion,
+      Azure::Core::Internal::Http::HttpPipeline(
+          options, "KeyVault", apiVersion, std::move(perRetrypolicies), {}));
 }
 
 Azure::Core::Response<KeyVaultKey> KeyClient::GetKey(
@@ -75,6 +67,51 @@ Azure::Core::Response<KeyVaultKey> KeyClient::CreateKey(
         return Details::KeyVaultKeyDeserialize(name, rawResponse);
       },
       {Details::KeysPath, name, "create"});
+}
+
+Azure::Core::Response<KeyVaultKey> KeyClient::CreateEcKey(
+    CreateEcKeyOptions const& ecKeyOptions,
+    Azure::Core::Context const& context) const
+{
+  std::string const& keyName = ecKeyOptions.GetName();
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Post,
+      Details::KeyRequestParameters(ecKeyOptions),
+      [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
+        return Details::KeyVaultKeyDeserialize(keyName, rawResponse);
+      },
+      {Details::KeysPath, keyName, "create"});
+}
+
+Azure::Core::Response<KeyVaultKey> KeyClient::CreateRsaKey(
+    CreateRsaKeyOptions const& rsaKeyOptions,
+    Azure::Core::Context const& context) const
+{
+  std::string const& keyName = rsaKeyOptions.GetName();
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Post,
+      Details::KeyRequestParameters(rsaKeyOptions),
+      [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
+        return Details::KeyVaultKeyDeserialize(keyName, rawResponse);
+      },
+      {Details::KeysPath, keyName, "create"});
+}
+
+Azure::Core::Response<KeyVaultKey> KeyClient::CreateOctKey(
+    CreateOctKeyOptions const& octKeyOptions,
+    Azure::Core::Context const& context) const
+{
+  std::string const& keyName = octKeyOptions.GetName();
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Post,
+      Details::KeyRequestParameters(octKeyOptions),
+      [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
+        return Details::KeyVaultKeyDeserialize(keyName, rawResponse);
+      },
+      {Details::KeysPath, keyName, "create"});
 }
 
 Azure::Security::KeyVault::Keys::DeleteKeyOperation KeyClient::StartDeleteKey(