+++ sdk/storage/azure-storage-blobs/src/blob_service_client.cpp
@@ -36,7 +36,7 @@ namespace Azure { namespace Storage { namespace Blobs {
       const BlobClientOptions& options)
       : m_serviceUrl(serviceUrl)
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(
         Storage::Details::ConstructPolicies(
@@ -54,7 +54,7 @@ namespace Azure { namespace Storage { namespace Blobs {
       const BlobClientOptions& options)
       : m_serviceUrl(serviceUrl)
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     Azure::Core::Http::TokenRequestOptions tokenOptions;
     tokenOptions.Scopes.emplace_back(Storage::Details::StorageScope);
@@ -74,7 +74,7 @@ namespace Azure { namespace Storage { namespace Blobs {
       const BlobClientOptions& options)
       : m_serviceUrl(serviceUrl)
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(
         Storage::Details::ConstructPolicies(