+++ sdk/core/azure-core/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.7 (Unreleased)
 
+### New Features
+- Added `HttpPolicyOrder` for adding custom Http policies to sdk clients.
+
 ### Breaking Changes
 
 - Removed `Azure::Core::Http::HttpPipeline` by making it internal, used only within the SDK.
@@ -15,6 +18,10 @@
 - Removed `LimitBodyStream`.
 - Renamed `AddHeader()` from `Request` and `RawResponse` to `SetHeader()`.
 - Introduced `Azure::Core::CaseInsensitiveMap` which is now used to store headers in `Azure::Core::Http::Request` and `Azure::Core::Http::RawResponse`.
+- Renamed `TransportPolicyOptions` to `TransportOptions`.
+- Renamed `TelemetryPolicyOptions` to `TelemetryOptions`.
+- Renamed `ValuePolicyOptions` to `ValueOptions`.
+- Removed `StartTry()` from `Azure::Core::Http::Request`.
 
 ### Bug Fixes